Inject some precompiler def to override try/catch (workaround)
Joern v2.0.321 fails to import the source code if a single file
contains a try/catch statement. To workaround this issue we inject
some precompiler instruction to turn in noop the try/catch.
Thanksfully, joern understands and processes these instructions
so from joern's perspective there is no try/catch and the bug
is not triggered.